+++ tests/testthat/test-template.R
@@ -33,3 +33,6 @@ test_that("moderncv", {
 test_that("latexcv", {
   expect_knit("latexcv")
 })
+test_that("markdowncv", {
+  expect_knit("markdowncv")
+})